setValue handles select fields
Not sure how people feel about this one, but in my specific case at the time I call `setValue`, I don't know the element type.  It would be nice, since the driver does know, if the `setValue` function noticed if it was of type `select` and pass it on to the `selectOption` function.  Currently, it doesn't fail, it just doesn't work.