Search for options within parent
When searching for option elements, we must limit the scope of the find
to the select that they belong to.

Where multiple select fields exist on the page, and they have options
children with the same value, the first one on the page was being
selected rather than the intended one.